Use size_t not int for InboxTraffic

+++ source/DataHelper.h
@@ -350,7 +350,7 @@ private:
       H5::DataSpace(2,dims)
     );
 
-    int data[dw.GetSize()];
+    size_t data[dw.GetSize()];
 
     for (size_t i = 0; i < dw.GetSize(); ++i) {
       data[i] = dw.man->Inbox(i).GetTraffic(dir);
